+++ lib/styles/style_10_bottom_navbar.dart
@@ -5,11 +5,13 @@ class Style10BottomNavBar extends StatefulWidget {
     required this.navBarConfig,
     this.navBarDecoration = const NavBarDecoration(),
     this.itemAnimationProperties = const ItemAnimation(),
+    this.height = kBottomNavigationBarHeight,
     super.key,
   });
 
   final NavBarConfig navBarConfig;
   final NavBarDecoration navBarDecoration;
+  final double height;
 
   /// This controls the animation properties of the items of the NavBar.
   final ItemAnimation itemAnimationProperties;
@@ -41,7 +43,7 @@ class _Style10BottomNavBarState extends State<Style10BottomNavBar>
       );
       _animationList.add(
         Tween(
-          begin: Offset(0, widget.navBarConfig.navBarHeight),
+          begin: Offset(0, widget.height),
           end: Offset.zero,
         )
             .chain(CurveTween(curve: widget.itemAnimationProperties.curve))
@@ -107,7 +109,7 @@ class _Style10BottomNavBarState extends State<Style10BottomNavBar>
     }
     return DecoratedNavBar(
       decoration: widget.navBarDecoration,
-      height: widget.navBarConfig.navBarHeight,
+      height: widget.height,
       child: Row(
         mainAxisAlignment: MainAxisAlignment.spaceAround,
         children: widget.navBarConfig.items.map((item) {